anchor_lang

Module accounts

Account types that can be used in the account validation struct.

Modules

account

Account container that checks ownership on deserialization.

account_info

AccountInfo can be used as a type but Unchecked Account should be used instead.

account_loader

Type facilitating on demand zero copy deserialization.

boxed

Box type to save stack space.

interface

Type validating that the account is one of a set of given Programs

interface_account

Account container that checks ownership on deserialization.

option

Option type for optional accounts.

program

Type validating that the account is the given Program

signer

Type validating that the account signed the transaction

system_account

Type validating that the account is owned by the system program

sysvar

Type validating that the account is a sysvar and deserializing it

unchecked_account

Explicit wrapper for AccountInfo types to emphasize that no checks are performed