Struct anchor_lang::accounts::account_loader::AccountLoader

pub struct AccountLoader<'info, T: ZeroCopy + Owner> { /* private fields */ }

Type facilitating on demand zero copy deserialization.

Note that using accounts in this way is distinctly different from using, for example, the Account. Namely, one must call

load_init after initializing an account (this will ignore the missing account discriminator that gets added only after the user’s instruction code)

load when the account is not mutable

load_mut when the account is mutable

For more details on zero-copy-deserialization, see the account attribute.

⚠️ When using this type it's important to be mindful of any calls to the load functions so as not to induce a RefCell panic, especially when sharing accounts across CPI boundaries. When in doubt, one should make sure all refs resulting from a call to a load function are dropped before CPI. This can be done explicitly by calling drop(my_var) or implicitly by wrapping the code using the Ref in braces {..} or moving it into its own function.

Example

use anchor_lang::prelude::*;

declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");

#[program]
pub mod bar {
    use super::*;

    pub fn create_bar(ctx: Context<CreateBar>, data: u64) -> Result<()> {
        let bar = &mut ctx.accounts.bar.load_init()?;
        bar.authority = ctx.accounts.authority.key();
        bar.data = data;
        Ok(())
    }

    pub fn update_bar(ctx: Context<UpdateBar>, data: u64) -> Result<()> {
        (*ctx.accounts.bar.load_mut()?).data = data;
        Ok(())
    }
}

#[account(zero_copy)]
#[derive(Default)]
pub struct Bar {
    authority: Pubkey,
    data: u64
}

#[derive(Accounts)]
pub struct CreateBar<'info> {
    #[account(
        init,
        payer = authority
    )]
    bar: AccountLoader<'info, Bar>,
    #[account(mut)]
    authority: Signer<'info>,
    system_program: AccountInfo<'info>,
}

#[derive(Accounts)]
pub struct UpdateBar<'info> {
    #[account(
        mut,
        has_one = authority,
    )]
    pub bar: AccountLoader<'info, Bar>,
    pub authority: Signer<'info>,
}

Implementations

impl<'info, T: ZeroCopy + Owner> AccountLoader<'info, T>

pub fn try_from(
acc_info: &AccountInfo<'info>
) -> Result<AccountLoader<'info, T>>

Constructs a new Loader from a previously initialized account.

pub fn try_from_unchecked(
_program_id: &Pubkey,
acc_info: &AccountInfo<'info>
) -> Result<AccountLoader<'info, T>>

Constructs a new Loader from an uninitialized account.

pub fn load(&self) -> Result<Ref<'_, T>>

Returns a Ref to the account data structure for reading.

pub fn load_mut(&self) -> Result<RefMut<'_, T>>

Returns a RefMut to the account data structure for reading or writing.

pub fn load_init(&self) -> Result<RefMut<'_, T>>

Returns a RefMut to the account data structure for reading or writing. Should only be called once, when the account is being initialized.