anchor_lang

Trait AccountsClose

pub trait AccountsClose<'info>: ToAccountInfos<'info> {
    // Required method
    fn close(&self, sol_destination: AccountInfo<'info>) -> Result<()>;
}

The close procedure to initiate garabage collection of an account, allowing one to retrieve the rent exemption.

Required Methods

fn close(&self, sol_destination: AccountInfo<'info>) -> Result<()>