Trait anchor_lang::ToAccountMetas

pub trait ToAccountMetas {
    fn to_account_metas(&self, is_signer: Option<bool>) -> Vec<AccountMeta>;
}

Transformation to AccountMeta structs.

Required Methods

fn to_account_metas(&self, is_signer: Option<bool>) -> Vec<AccountMeta>

is_signer is given as an optional override for the signer meta field. This covers the edge case when a program-derived-address needs to relay a transaction from a client to another program but sign the transaction before the relay. The client cannot mark the field as a signer, and so we have to override the is_signer meta field given by the client.